Moved create Dict back to letters game
