+++ LettersGame/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["CountdownSolver"]
+__all__ = ["CountdownSolver", "CreateDict"]
